requirements.txt
nextcord==2.4.2
Cybernator==0.6.3
PyMySQL==1.0.3
pypresence==4.2.1
psycopg2==2.9.6
openai==0.27.4
Pillow==9.3.0
aiohttp==3.8.4
function-cooldowns==1.7.0